[droidcamsrc] call droid_media_camera_take_picture from another thread
take_picture call in HAL sometimes need to read information from preview
frames for e.g. auto exposure when using flash. If the calling thread
also handles pulling preview frames, this will create a deadlock.

To prevent this, I've moved the actual droid_media_camera_take_picture()
call to a function that will be called in a background thread. The
GstTaskPool is used to execute the function in the glib's thread pool,
so that the new thread doesn't have to be created every time.

This fixed taking a picture with the flash on on Fairphone 2 with
Halium 7.1-based port and Ubuntu Touch.